requirements.txt
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.0
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
marshmallow==3.22.0
networkx==3.4.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
pytest==8.3.3
pytest-asyncio==0.25.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
six==1.17.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
Wikipedia-API==0.8.1
